+++ fast-ftrl-proximal/src/main/java/io/scaledml/ftrl/FtrlOptions.java
@@ -1,4 +1,4 @@
-package io.scaledml;
+package io.scaledml.ftrl;
 
 import com.lexicalscope.jewel.cli.Option;
