+++ fast-ftrl-proximal/src/main/java/io/scaledml/ftrl/inputformats/InputFormat.java
@@ -1,0 +1,8 @@
+package io.scaledml.ftrl.inputformats;
+
+import io.scaledml.ftrl.SparseItem;
+import io.scaledml.ftrl.io.LineBytesBuffer;
+
+public interface InputFormat {
+    void parse(LineBytesBuffer line, SparseItem item);
+}
