+++ fast-ftrl-proximal/src/main/java/io/scaledml/ftrl/io/LineBytesBuffer.java
@@ -1,4 +1,4 @@
-package io.scaledml.io;
+package io.scaledml.ftrl.io;
 
 import com.google.common.base.Charsets;
 import it.unimi.dsi.fastutil.bytes.ByteArrays;
