+++ fast-ftrl-proximal/src/main/java/io/scaledml/ftrl/outputformats/CollectStatisticsOutputFormat.java
@@ -1,7 +1,7 @@
-package io.scaledml.outputformats;
+package io.scaledml.ftrl.outputformats;
 
 
-import io.scaledml.SparseItem;
+import io.scaledml.ftrl.SparseItem;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
