+++ fast-ftrl-proximal/src/main/java/io/scaledml/ftrl/outputformats/OutputFormat.java
@@ -1,6 +1,6 @@
-package io.scaledml.outputformats;
+package io.scaledml.ftrl.outputformats;
 
-import io.scaledml.SparseItem;
+import io.scaledml.ftrl.SparseItem;
 
 import java.io.Closeable;
