--- fast-ftrl-proximal/src/main/java/io/scaledml/inputformats/InputFormat.java
@@ -1,8 +1,0 @@
-package io.scaledml.inputformats;
-
-import io.scaledml.SparseItem;
-import io.scaledml.io.LineBytesBuffer;
-
-public interface InputFormat {
-    void parse(LineBytesBuffer line, SparseItem item);
-}
