+++ fast-ftrl-proximal/src/test/java/io/scaledml/ftrl/VowpalWabbitFormatTest.java
@@ -1,9 +1,9 @@
-package io.scaledml;
+package io.scaledml.ftrl;
 
 import com.google.common.base.Charsets;
-import io.scaledml.inputformats.InputFormat;
-import io.scaledml.inputformats.VowpalWabbitFormat;
-import io.scaledml.io.LineBytesBuffer;
+import io.scaledml.ftrl.inputformats.VowpalWabbitFormat;
+import io.scaledml.ftrl.inputformats.InputFormat;
+import io.scaledml.ftrl.io.LineBytesBuffer;
 import it.unimi.dsi.fastutil.io.FastBufferedInputStream;
 import org.apache.commons.io.input.ReaderInputStream;
 import org.junit.Test;
