+++ fast-ftrl-proximal/src/test/java/io/scaledml/ftrl/io/LineBytesBufferTest.java
@@ -1,4 +1,4 @@
-package io.scaledml.io;
+package io.scaledml.ftrl.io;
 
 
 import com.google.common.base.Charsets;
